requirements.txt
yt-dlp
litellm
pyyaml
rich
openai
faster-whisper